DRFAuthenticationApp/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==38.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-extensions==3.2.0
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.13.1
Faker==14.2.0
flake8==5.0.4
idna==3.3
Markdown==3.4.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-Pillow==9.5.0.0
types-requests==2.28.11.17
types-urllib3==1.26.25.10
typing_extensions==4.4.0
urllib3==1.26.11
wrapt==1.14.1
